#245 Refactored code
