+++ test/LightBDD.Framework.UnitTests/AssemblyInfo.cs
@@ -1,3 +1,5 @@
 ﻿using NUnit.Framework;
 
 [assembly: Parallelizable(ParallelScope.Fixtures)]
+[assembly: SetCulture("en-us")]
+[assembly: SetUICulture("en-us")]
